Refactors Helpers (#1031).
