+++ spec/helpers/exterior_links_helper_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require 'rails_helper'
 
-RSpec.describe UsersHelper, type: :helper do
+RSpec.describe ExteriorLinksHelper, type: :helper do
   around do |example|
     ENV['ALMA'] = 'example'
     ENV['INSTITUTION'] = '01_Example'
